Xamarin: LaunchMode to singleinstance and move serial init
